Background class added, doors now seem to work properly

+++ AdventureGame/AdventureGame/Classes/Input and output (non-visual)/SaveHandler.cs
@@ -55,7 +55,8 @@ namespace AdventureGame
             string[] filePaths = Directory.GetFiles(CurrentSavePath);
             foreach (string filePath in filePaths)
             {
-                DeleteFile(filePath);
+                if(filePath.EndsWith(".sav"))
+                    DeleteFile(filePath);
             }
         }
